+++ src/index.ts
@@ -1,4 +1,4 @@
-export const BPD_TOOLKIT_VERSION = "0.1.10";
+export const BPD_TOOLKIT_VERSION = "0.1.11";
 /**
  * Checks if value is undefined
  * @param val value