+++ changelog.md
@@ -1,10 +1,10 @@
 # bpd toolkit
-# [0.1.10] WIP
+# [0.1.11] 2021-02-16 
 * [Added] Task to handle async operations
 * [Added] Copy text
 * [Added] Generate GUID, isFullscreen
 * [Added] makeArray
-# [0.1.9] 2020-01-10
+# [0.1.9] 2021-01-10
 * [Added] **getLocation** method
 * [Added] **where** and **findFirst** collection methods
 # [0.1.8] 2020-12-26